% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_parse.R
\name{addr_parse}
\alias{addr_parse}
\title{addr_parse}
\usage{
addr_parse(address.string, python.system.location = system("whereis python"),
  cole = FALSE)
}
\arguments{
\item{address.string}{a string to be parsed for address components}

\item{python.system.location}{string for filepath to python installation to
use for addr_parse.py}

\item{cole}{logical, convenience argument to set my python location on my local machine}

\item{return.call}{logical, return the call along with the parsed address?}
}
\value{
a data.frame with the address components as columns and optionally
  the submitted address string
}
\description{
This function utilizes a python script to parse out individual address
components from a string.
}
\details{
An address string is parsed for the following components: - `AddressNumber`,
`StreetName`, `StreetPostType`, `PlaceName`, `StateName`, `ZipCode`,
`StreetNamePreType`, `StreetNamePreDirectional`, `AddressNumberSuffix`,
`StreetNamePostDirectional`.  If any of these components are not present in
the string `NA` is returned for that component.
}
\examples{
addr_parse('3333 Burnet Ave, Cincinnati, OH 45229',cole=TRUE)
addr_parse('737 US 50 Cincinnati OH 45150',cole=TRUE)
addr_parse('3333 Burnet Ave, Cincinnati, OH 45229',
python.system.location='/Library/Frameworks/Python.framework/Versions/2.7/bin/python')
}

