% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_parse.R
\name{addr_parse}
\alias{addr_parse}
\title{Parse an address string into individual components.}
\usage{
addr_parse(address.string)
}
\arguments{
\item{address.string}{a string to be parsed for address components}
}
\value{
a data.frame with the address components
}
\description{
This function utilizes a python script to parse out individual address
components from a string.
}
\details{
An address string is parsed for the following components: - `AddressNumber`,
`StreetName`, `StreetPostType`, `PlaceName`, `StateName`, `ZipCode`,
`StreetNamePreType`, `StreetNamePreDirectional`, `AddressNumberSuffix`,
`StreetNamePostDirectional`.  If any of these components are not present in
the string `NA` is returned for that component.

This function depends on having a sufficient python binary installed and
having the module `usaddress` available. Use `pip install usaddress` to
install this module after python is installed.
}
\examples{
addr_parse('3333 Burnet Ave, Cincinnati, OH 45229')
addr_parse('737 US 50 Cincinnati OH 45150')
}

