% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeCAGIS.R
\name{geocodeCAGIS}
\alias{geocodeCAGIS}
\title{Geocode Cincinnati, OH area address using offline exact location files}
\usage{
geocodeCAGIS(addr_string, verbose = FALSE, return.call = FALSE)
}
\arguments{
\item{verbose}{logical, return method and matching score?}

\item{address_string}{a single string that will be geocoded}
}
\value{
data.frame with lat/lon coords and optionally method score
}
\description{
Geocode an address using offline shapefile from CAGIS. Requires a
\code{R/sysdata.rda} file, see details for more information.
}
\details{
This function parses a given address string into address components and
attempts to match the address to CAGIS data.  The best match is returned and
the score represents how many insertions/deletions/rearrangements were needed
to match the input string to the address data.

The \code{sysdata.rda} file is available via personal communication (cole dot
brokamp at gmail dot com). Alternatively, build the system data file on your
own, using updated CAGIS files or a different vintage of TIGER/Line files.
See \code{README.md} for details on this operation. Make sure to `R CMD BATCH
--vanilla make_sysdata.R` and rebuild the package if using your own address
reference files.

This function will return NA if the zip code of the address string
does not begin with 450, 451, or 452.
}
\examples{
geocodeCAGIS('3333 Burnet Ave, Cincinnati, OH 45229')
geocodeCAGIS('3333 Burnet Ave, Cincinnati, OH 45229',verbose=TRUE,return.call=TRUE)
geocodeCAGIS('1456 Main St. 23566')
}

